+++ app/src/main/java/com/polishmediahub/app/data/plugin/models/InstallablePlugin.kt
@@ -32,6 +32,7 @@ sealed class InstallablePlugin {
         override val fileSize: Long? = null,
         val pkg: String = "",
         val lang: String = "en",
-        val nsfw: Boolean = false
+        val nsfw: Boolean = false,
+        val mainClass: String = ""
     ) : InstallablePlugin()
 }